Compute definition on the fly for local symbols.
SemanticDB v3 changed the encoding for local symbols so that they are
only unique within a single source file but can have conflicts with
local symbols in other source files. This commit refactors the indexing
pipeline to not index local symbols and compute symbol data for local
symbols on the fly instead. This refactoring should lower memory usage
as well.